完成新增用户api

+++ src/server/model/user.go
@@ -13,10 +13,10 @@ type User struct {
 	Password  string `json:"password"`  //密码
 	Telephone string `json:"telephone"` //服务电话
 	Email     string `json:"email"`     //邮箱
-	ParentId  uint32 `json:"parent_id"` //父id
-	Gender    uint16 `json:"gender"`    //性别
-	Age       uint16 `json:"age"`       //年龄
-	Status    uint16 `json:"status"`    //状态
+	ParentId  int    `json:"parent_id"` //父id
+	Gender    int    `json:"gender"`    //性别
+	Age       int    `json:"age"`       //年龄
+	Status    int    `json:"status"`    //状态
 }
 
 func (User) TableName() string {
